requirements.txt
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
PyYAML==6.0
PyYAML==6.0.2
scikit_learn==1.2.0
scipy==1.9.3
torch==1.13.1
tqdm==4.64.1
